requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bson==0.5.8
dataclasses==0.6
Django==2.2.13
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-credit-cards==0.4.1
# django-suit
https://github.com/darklow/django-suit/tarball/v2
djongo==1.3.1
dnspython==1.16.0
Pillow==7.1.2
pymongo==3.7.2
python-dateutil==2.8.1
pytz==2018.7
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
